# openapi/openapi.json
{
  "components": {
    "schemas": {
...
      "BreathingAudioMigrationResult": {
        "properties": {
          "dryRun": {
            "type": "boolean"
          },
          "failed": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "moved": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "skipped": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "targetBucket": {
            "type": "string"
          }
        },
        "type": "object"
      },
      "BreathingCreateRequest": {
        "properties": {
          "color": {
            "minLength": 1,
            "type": "string"
          },
          "cycles": {
            "format": "int32",
            "minimum": 1,
            "type": "integer"
          },
          "description": {
            "type": "string"
          },
          "displayOrder": {
            "format": "int32",
            "type": "integer"
          },
          "isPremium": {
            "type": "boolean"
          },
          "name": {
            "minLength": 1,
            "type": "string"
          },
          "phases": {
            "items": {
              "$ref": "#/components/schemas/BreathingPhaseCreateRequest"
            },
            "minItems": 1,
            "type": "array"
          }
        },
        "required": [
          "cycles",
          "phases"
        ],
        "type": "object"
      },
      "BreathingDto": {
        "properties": {
          "color": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "cycles": {
            "format": "int32",
            "type": "integer"
          },
          "description": {
            "type": "string"
          },
          "displayOrder": {
            "format": "int32",
            "type": "integer"
          },
          "icon": {
            "type": "string"
          },
          "id": {
            "format": "int64",
            "type": "integer"
          },
          "isPremium": {
            "type": "boolean"
          },
          "locked": {
            "type": "boolean"
          },
          "name": {
            "type": "string"
          },
          "phases": {
            "items": {
              "$ref": "#/components/schemas/BreathingPhaseDto"
            },
            "type": "array"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "type": "object"
      },
      "BreathingPhaseAudioDto": {
        "properties": {
          "id": {
            "format": "int64",
            "type": "integer"
          },
          "url": {
            "type": "string"
          }
        },
        "type": "object"
      },
      "BreathingPhaseCreateRequest": {
        "properties": {
          "audioObjectIdsToRemove": {
            "items": {
              "format": "int64",
              "type": "integer"
            },
            "type": "array"
          },
          "color": {
            "minLength": 1,
            "type": "string"
          },
          "displayOrder": {
            "format": "int32",
            "type": "integer"
          },
          "durationSeconds": {
            "format": "int32",
            "minimum": 1,
            "type": "integer"
          },
          "id": {
            "format": "int64",
            "type": "integer"
          },
          "label": {
            "minLength": 1,
            "type": "string"
          },
          "name": {
            "minLength": 1,
            "type": "string"
          },
          "newAudioFileIndices": {
            "items": {
              "format": "int32",
              "type": "integer"
            },
            "type": "array"
          }
        },
        "required": [
          "durationSeconds"
        ],
        "type": "object"
      },
      "BreathingPhaseDto": {
        "properties": {
          "audioFiles": {
            "items": {
              "$ref": "#/components/schemas/BreathingPhaseAudioDto"
            },
            "type": "array"
          },
          "color": {
            "type": "string"
          },
          "displayOrder": {
            "format": "int32",
            "type": "integer"
          },
          "duration": {
            "format": "int32",
            "type": "integer"
          },
          "id": {
            "format": "int64",
            "type": "integer"
          },
          "label": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        },
        "type": "object"
      },
      "BreathingReorderItem": {
        "properties": {
          "displayOrder": {
            "format": "int32",
            "type": "integer"
          },
          "id": {
            "format": "int64",
            "type": "integer"
          }
        },
        "required": [
          "displayOrder",
          "id"
        ],
        "type": "object"
      },
      "BreathingUpdateRequest": {
        "properties": {
          "color": {
            "type": "string"
          },
          "cycles": {
            "format": "int32",
            "minimum": 1,
            "type": "integer"
          },
          "description": {
            "type": "string"
          },
          "displayOrder": {
            "format": "int32",
            "type": "integer"
          },
          "isPremium": {
            "type": "boolean"
          },
          "name": {
            "type": "string"
          },
          "phases": {
            "items": {
              "$ref": "#/components/schemas/BreathingPhaseCreateRequest"
            },
            "type": "array"
          }
        },
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "bearerFormat": "JWT",
        "description": "Keycloak-issued access token. Two realms are accepted, the mobile realm and the admin realm; the token's issuer selects which. Endpoints under /admin additionally require the ADMIN role.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {
      "name": "Yogida Team",
      "url": "https://github.com/yogida"
    },
    "description": "REST API for the Yogida meditation platform. \nProvides presigned Cloudflare R2 streaming URLs and a catalog of available meditation media grouped by bucket.",
    "title": "Meditation Service API",
    "version": "v1"
  },
  "openapi": "3.1.0",
  "paths": {
    "/admin/breathing": {
      "get": {
        "operationId": "listAdminBreathingExercises",
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/BreathingDto"
                  },
                  "type": "array"
                }
              }
            },
            "description": "Exercises retrieved"
          }
        },
        "summary": "List all breathing exercises",
        "tags": [
          "Admin — Breathing"
        ]
      },
      "post": {
        "description": "Creates a new exercise. Send as multipart/form-data: part 'meta' (JSON, BreathingCreateRequest) + part 'iconFile' (required image). Audio files are added separately via POST /admin/breathing/phases/{phaseId}/audio.",
        "operationId": "createBreathingExercise",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "properties": {
                  "audioFiles": {
                    "items": {
                      "format": "binary",
                      "type": "string"
                    },
                    "type": "array"
                  },
                  "iconFile": {
                    "format": "binary",
                    "type": "string"
                  },
                  "meta": {
                    "$ref": "#/components/schemas/BreathingCreateRequest"
                  }
                },
                "required": [
                  "iconFile",
                  "meta"
                ],
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "201": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BreathingDto"
                }
              }
            },
            "description": "Exercise created"
          },
          "400": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BreathingDto"
                }
              }
            },
            "description": "Validation error"
          }
        },
        "summary": "Create a breathing exercise",
        "tags": [
          "Admin — Breathing"
        ]
      }
    },
    "/admin/breathing/audio/migrate-to-private": {
      "post": {
        "description": "One-off migration. Breathing phase audio was originally uploaded to the PUBLIC bucket and served by unsigned URL, so premium exercises were readable by anyone holding the link. New uploads now go to the private bucket; this moves the objects that predate that change and repoints their s3_object rows. Idempotent and safe to re-run. Run it with dryRun=true first — the delete of the public copy is not reversible.",
        "operationId": "migrateBreathingAudioToPrivateBucket",
        "parameters": [
          {
            "description": "Report what would move without changing anything. Defaults to true.",
            "in": "query",
            "name": "dryRun",
            "required": false,
            "schema": {
              "default": true,
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BreathingAudioMigrationResult"
                }
              }
            },
            "description": "Migration report"
          },
          "403": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BreathingAudioMigrationResult"
                }
              }
            },
            "description": "Caller is not an administrator"
          },
          "500": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BreathingAudioMigrationResult"
                }
              }
            },
            "description": "Private bucket is not configured"
          }
        },
        "summary": "Move breathing phase audio into the private bucket",
        "tags": [
          "Admin — Breathing"
        ]
      }
    },
    "/admin/breathing/phases/{phaseId}/audio": {
      "post": {
        "description": "Uploads one audio file and links it to the specified phase. Multiple calls add multiple audio options; the mobile client picks one at random.",
        "operationId": "addBreathingPhaseAudio",
        "parameters": [
          {
            "description": "Phase ID",
            "in": "path",
            "name": "phaseId",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "properties": {
                  "audioFile": {
                    "format": "binary",
                    "type": "string"
                  }
                },
                "required": [
                  "audioFile"
                ],
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BreathingDto"
                }
              }
            },
            "description": "Audio added; returns updated exercise"
          },
          "404": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BreathingDto"
                }
              }
            },
            "description": "Phase not found"
          }
        },
        "summary": "Upload an audio file to a phase",
        "tags": [
          "Admin — Breathing"
        ]
      }
    },
    "/admin/breathing/phases/{phaseId}/audio/{audioObjectId}": {
      "delete": {
        "operationId": "removeBreathingPhaseAudio",
        "parameters": [
          {
            "description": "Phase ID",
            "in": "path",
            "name": "phaseId",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          },
          {
            "description": "S3 audio object ID to remove",
            "in": "path",
            "name": "audioObjectId",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BreathingDto"
                }
              }
            },
            "description": "Audio removed; returns updated exercise"
          },
          "404": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BreathingDto"
                }
              }
            },
            "description": "Phase or audio object not found"
          }
        },
        "summary": "Remove an audio file from a phase",
        "tags": [
          "Admin — Breathing"
        ]
      }
    },
    "/admin/breathing/reorder": {
      "put": {
        "description": "Bulk-updates the displayOrder of exercises. Send the full ordered list with new positions.",
        "operationId": "reorderBreathingExercises",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "items": {
                  "$ref": "#/components/schemas/BreathingReorderItem"
                },
                "type": "array"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Reorder applied"
          }
        },
        "summary": "Reorder breathing exercises",
        "tags": [
          "Admin — Breathing"
        ]
      }
    },
    "/admin/breathing/{id}": {
      "delete": {
        "operationId": "deleteBreathingExercise",
        "parameters": [
          {
            "description": "Breathing exercise ID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Exercise deleted"
          },
          "404": {
            "description": "Exercise not found"
          }
        },
        "summary": "Delete a breathing exercise",
        "tags": [
          "Admin — Breathing"
        ]
      },
      "get": {
        "operationId": "getAdminBreathingExerciseById",
        "parameters": [
          {
            "description": "Breathing exercise ID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BreathingDto"
                }
              }
            },
            "description": "Exercise found"
          },
          "404": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BreathingDto"
                }
              }
            },
            "description": "Exercise not found"
          }
        },
        "summary": "Get a breathing exercise by ID",
        "tags": [
          "Admin — Breathing"
        ]
      },
      "put": {
        "description": "Partially updates the exercise. Send as multipart/form-data: part 'meta' (JSON, BreathingUpdateRequest) + optional part 'iconFile' (replaces current icon). If phases are provided they fully replace the current phase list.",
        "operationId": "updateBreathingExercise",
        "parameters": [
          {
            "description": "Breathing exercise ID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "properties": {
                  "audioFiles": {
                    "items": {
                      "format": "binary",
                      "type": "string"
                    },
                    "type": "array"
                  },
                  "iconFile": {
                    "format": "binary",
                    "type": "string"
                  },
                  "meta": {
                    "$ref": "#/components/schemas/BreathingUpdateRequest"
                  }
                },
                "required": [
                  "meta"
                ],
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BreathingDto"
                }
              }
            },
            "description": "Exercise updated"
          },
          "404": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BreathingDto"
                }
              }
            },
            "description": "Exercise not found"
          }
        },
        "summary": "Update a breathing exercise",
        "tags": [
          "Admin — Breathing"
        ]
      }
    },
...
    "/breathing/{id}": {
      "get": {
        "operationId": "getBreathingExerciseById",
        "parameters": [
          {
            "description": "Breathing exercise ID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BreathingDto"
                }
              }
            },
            "description": "Exercise found"
          },
          "404": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BreathingDto"
                }
              }
            },
            "description": "Exercise not found"
          }
        },
        "summary": "Get a breathing exercise by ID",
        "tags": [
          "Breathing"
        ]
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "servers": [
    {
      "description": "Current environment",
      "url": "/api"
    }
  ],
...
}